fix: Hilt dependency injection implementation
Even though Hilt was added as a dependency and, at places, the `@Inject` annotation was being used, the injection was not happening because there was no Hilt application and entry point set up which are required for the injection to happen.
Switched from KAPT to KSP which analyzes Kotlin code directly and improves build time by making it 2x faster.