fix: Infinite loading state caused by recomposition
The `PostViewModel` being instantiated as a local variable inside the `PostScreen` composable was not being preserved during recompositions, which caused it to get instantiated at each recomposition.
This, coupled with the call to `loadPost` function in the `PostView` composable, caused the `posts` state to change at each recomposition, causing yet another recomposition, producing a never-ending loop.

To address the issue, I have moved the `PostViewModel` to the Navigation composable and used the Compose ViewModel Lifecycle function `viewModel()` to obtain an instance of the `PostViewModel`, which is not recreated during recompositions.
Also, the `PostViewModel` loads the posts when it is instantiated, so posts are not loaded again during each recomposition.