perf: Lazy load posts using `LazyColumn`
Using `Column` means that all the posts are loaded into memory at once. This is not scalable and is terrible in terms of performance. It can be observed in the debugger that whenever we change the screen orientation, a log prints out that tells us that the app skipped some frames and that it might be doing a lot of work on the main thread. This is because all the children of the `Column` are loaded at once at each configuration change.
We should instead use `LazyColumn` which recycles the composables and loads only the children that are inside the viewport. With this change, there will be no frames getting skipped.