chore: Repository interface & di
This commit introduces an abstraction for the post repository. This allows us to provide a mock implementation of the repository so we can test it.
Check the `PostRepositoryTest` file.

`AppModule` allows us to let Hilt know how to generate the dependencies required by our classes. We delegate the job of creating the `PostService` to Hilt so we do not need to worry about building the Retrofit client.
We are also making Hilt generate the `PostRepository` class for us. Whenever we need to swap to repository to something else, all we would need to is return the new repository implementation class and it will be used across our app.